feat : upgrade to spring boot 3.4.0-RC1
