+++ jpa/boot-data-multipledatasources/src/main/java/com/example/multipledatasources/entities/cardholder/CardHolder.java
@@ -1,10 +1,11 @@
-package com.example.multipledatasources.model.member;
+package com.example.multipledatasources.entities.cardholder;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import java.util.Objects;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -17,24 +18,25 @@ import org.hibernate.Hibernate;
 @Setter
 @ToString
 @RequiredArgsConstructor
-public class Member {
+@Table(name = "card_holder")
+public class CardHolder {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-    @Column(nullable = false)
-    private String name;
 
     @Column(nullable = false, name = "member_id")
     private String memberId;
+
+    @Column(nullable = false, name = "card_number")
+    private String cardNumber;
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
-        Member member = (Member) o;
-        return id != null && Objects.equals(id, member.id);
+        CardHolder that = (CardHolder) o;
+        return id != null && Objects.equals(id, that.id);
     }
 
     @Override
