+++ jpa/boot-data-multipledatasources/src/main/java/com/example/multipledatasources/entities/member/Member.java
@@ -1,11 +1,10 @@
-package com.example.multipledatasources.model.cardholder;
+package com.example.multipledatasources.entities.member;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.Table;
 import java.util.Objects;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -18,25 +17,24 @@ import org.hibernate.Hibernate;
 @Setter
 @ToString
 @RequiredArgsConstructor
-@Table(name = "card_holder")
-public class CardHolder {
+public class Member {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @GeneratedValue(strategy = GenerationType.SEQUENCE)
     private Long id;
+
+    @Column(nullable = false)
+    private String name;
 
     @Column(nullable = false, name = "member_id")
     private String memberId;
-
-    @Column(nullable = false, name = "card_number")
-    private String cardNumber;
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
-        CardHolder that = (CardHolder) o;
-        return id != null && Objects.equals(id, that.id);
+        Member member = (Member) o;
+        return id != null && Objects.equals(id, member.id);
     }
 
     @Override
