+++ jpa/boot-data-multipledatasources/src/main/java/com/example/multipledatasources/repository/member/MemberRepository.java
@@ -1,6 +1,6 @@
 package com.example.multipledatasources.repository.member;
 
-import com.example.multipledatasources.model.member.Member;
+import com.example.multipledatasources.entities.member.Member;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
