+++ jpa/boot-data-multipledatasources/src/main/java/com/example/multipledatasources/service/impl/DetailsServiceImpl.java
@@ -1,8 +1,8 @@
 package com.example.multipledatasources.service.impl;
 
 import com.example.multipledatasources.dto.ResponseDto;
-import com.example.multipledatasources.model.cardholder.CardHolder;
-import com.example.multipledatasources.model.member.Member;
+import com.example.multipledatasources.entities.cardholder.CardHolder;
+import com.example.multipledatasources.entities.member.Member;
 import com.example.multipledatasources.repository.cardholder.CardHolderRepository;
 import com.example.multipledatasources.repository.member.MemberRepository;
 import com.example.multipledatasources.service.DetailsService;
