+++ jpa/boot-data-multipledatasources/src/test/java/com/example/multipledatasources/controller/DetailsControllerTest.java
@@ -10,8 +10,8 @@ import com.example.multipledatasources.service.DetailsService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 
 @WebMvcTest(controllers = DetailsController.class)
@@ -20,7 +20,7 @@ class DetailsControllerTest {
     @Autowired
     private MockMvc mockMvc;
 
-    @MockBean
+    @MockitoBean
     private DetailsService detailsService;
 
     @Test
